Made code more modular Kept types definitions under appropriate feature folders
