+++ src/components/uv_bar-chart/uv_bar-chart.types.ts
@@ -1,0 +1,15 @@
+import { UVItem } from '../../shared/Types';
+
+export type UVBarChartConfig = {
+  dimension ?: string;
+  categoryKey: string;
+  categoryShortKey: string;
+  valueAxis ?: object;
+  series ?: object;
+};
+
+export type UVBarChartProps = {
+  componentId: number;
+  config: UVBarChartConfig;
+  items: UVItem[];
+};
