+++ src/components/uv_number/uv_number.types.ts
@@ -1,0 +1,14 @@
+export type UVNumberProps = {
+  config ?: {
+    class ?: string;
+    title ?: {
+      class ?: string;
+    };
+    subtitle ?: {
+      class ?: string;
+    }
+  };
+  title: number;
+  label: string;
+  subtitle?: string;
+};
