+++ src/components/uv_pie/uv_pie.types.ts
@@ -1,0 +1,15 @@
+import { UVCategory } from '../../shared/Types';
+
+export type UVPieConfig = {
+  series: any;
+};
+
+export type UVPieData = {
+  categories: UVCategory[]
+};
+
+export type UVPieProps = {
+  componentId: number;
+  config: UVPieConfig;
+  categories: UVCategory[];
+};
