+++ src/modules/uv_dashboard/uv_dashboard.types.ts
@@ -1,0 +1,6 @@
+import { UVNumberProps } from '../../components/uv_number/uv_number.types';
+
+export type UVDashboardType = {
+  config ?: any;
+  uvNumbers: UVNumberProps[]
+};
